refactor(CompareBalls) : 클래스 이름 명사로 변경 (CompareBalls -> ComparisonBalls)
