+++ src/main/java/utils/ComparisonBalls.java
@@ -1,8 +1,8 @@
-package baseball;
+package utils;
 
 import static utils.Constant.*;
 
-public class CompareBalls {
+public class ComparisonBalls {
 
     public static int getStrikeCount(String computer, String player) {
         int strike = 0;
@@ -17,6 +17,7 @@ public class CompareBalls {
     public static int getBallCount(String computer, String player) {
 
         int ball = 0;
+        
         for (int i = 0; i < BALLS_LENGTH; i++) {
             if (computer.contains(String.valueOf(player.charAt(i)))) {
                 ball++;
